My Landscape Patch Here!

+++ src/tree_cmd.cpp
@@ -820,7 +820,7 @@ static void TileLoop_Trees(TileIndex tile)
 bool DecrementTreeCounter()
 {
 	/* Ensure _trees_tick_ctr can be decremented past zero only once for the largest map size. */
-	static_assert(2 * (MAX_MAP_SIZE_BITS - MIN_MAP_SIZE_BITS) - 4 <= std::numeric_limits<byte>::digits);
+	static_assert(2 * (MAX_MAP_SIZE_BITS - MIN_MAP_SIZE_BITS) - 4 <= std::numeric_limits<uint>::digits);
 
 	/* byte underflow */
 	byte old_trees_tick_ctr = _trees_tick_ctr;
